button group component for trade side
