--- web-front-end/src/CreateAccount/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateAccount";
