--- web-front-end/src/CreateAccountUser/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateAccountUser";
